Impr(sdk) Update Sekoia Automation Version
fix merge

cleaning

cleaning

Update Checkpoint/manifest.json


Cleaning

Fix tests

update mypy version

Try to fix black

Run black

remove

add missing package

Fix tests

Some fixes again

Run black

+++ Salesforce/client/schemas/token.py
@@ -20,7 +20,7 @@ class SalesforceToken(BaseModel):
     """Model to work with auth token with additional info."""
 
     token: HttpToken
-    created_at: int
+    created_at: float
     ttl: int
 
     def is_valid(self) -> bool:
